+++ NEWS.md
@@ -4,7 +4,14 @@ NEWS
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
 
-    #> Last Update: 2023-10-23 10:26:40.534476
+    #> Last Update: 2023-10-23 10:30:47.753788
+
+# Shiny_Test 0.0.0.9004
+
+Released - 2023-10-23
+
+- feat: Add links to files
+- refactor: Add global file
 
 # Shiny_Test 0.0.0.9003